+++ src/chartRenderer.ts
@@ -2,15 +2,32 @@ import { App } from "obsidian";
 import { parseCode } from "./utils/configParser";
 import { prepareDataVariables } from "./utils/prepareDataVariables";
 import * as echarts from 'echarts';
+import * as ecStat from 'echarts-stat';
 import type { RendererConfig } from "@hypersphere/sqlseal";
+import { ViewDefinition } from "@hypersphere/sqlseal/dist/src/grammar/parser";
+import { parseCodeAdvanced } from "./utils/advancedParser";
 
 interface Config {
     config: string
 }
 
+console.log((ecStat as any).transform)
+
+echarts.registerTransform((ecStat as any).transform.clustering);
+echarts.registerTransform((ecStat as any).transform.regression);
+echarts.registerTransform((ecStat as any).transform.histogram);
+
 export class ChartRenderer implements RendererConfig {
 
     constructor(private readonly app: App) {
+    }
+
+    get viewDefinition(): ViewDefinition {
+        return {
+            argument: 'javascriptTemplate',
+            name: 'chart',
+            singleLine: false
+        }
     }
 
     get rendererKey() {
@@ -18,22 +35,41 @@ export class ChartRenderer implements RendererConfig {
     }
 
     validateConfig(config: string): Config {
-        return { config }
+        return { config: config.trim() }
     }
 
     render(config: Config, el: HTMLElement) {
         let isRendered: boolean = false
         let chart: echarts.ECharts | null = null
         return {
-            render: ({ columns, data }: { columns: string[], data: Record<string, unknown>[]}) => {
+            render: ({ columns, data, flags }: { columns: string[], data: Record<string, unknown>[], flags: Record<string, boolean> }) => {
+
+                const isAdvancedMode = !!(flags?.isAdvancedMode)
+
+                if (config.config[0] !== '{' && !isAdvancedMode) {
+                    throw new Error('To process JavaScript, set ADVANCED MODE flag')
+                }
                 const { functions, variables } = prepareDataVariables({ columns, data })
-                const parsedConfig = parseCode(config.config, functions, variables)
+
+                let parsedConfig: Object = {}
+                if (isAdvancedMode) {
+                    try {
+                        parsedConfig = parseCodeAdvanced({ functions, variables }, config.config)
+                    } catch (e) {
+                        console.error(e)
+                        throw e
+                    }
+                } else {
+                    parsedConfig = parseCode(config.config, functions, variables) as Object
+                }
                 if (!parsedConfig || typeof parsedConfig !== 'object') {
                     throw new Error('Issue with parsing config')
                 }
                 const configRecord = parsedConfig as Record<string, any>
-                const dataset = [{ id: 'data', source: data }, ...(configRecord.dataset ?? [])]
-                configRecord.dataset = dataset
+                if (!configRecord.dataset) {
+                    const dataset = [{ id: 'data', source: data }, ...(configRecord.dataset ?? [])]
+                    configRecord.dataset = dataset
+                }
 
                 if (isRendered) {
                     // Data update