+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.2.0 (2025-03-15)
+feat: You can now use "ADVANCED MODE" to use full JavaScript to customise your charts
+feat: Enabling EChart-Stat module. You can now perform regression, clustering and histogram operations directly in Obsidian!
+
 # 0.1.1 (2025-02-01)
 chore: Fixed typing in configParser